Create Sales by Match
Alex works at a clothing store. There is a large pile of socks that must be paired by color for sale. Given an array of integers representing the color of each sock, determine how many pairs of socks with matching colors there are.

Return the total number of matching pairs of socks that Alex can sell.

Sample Input:

9
10 20 20 10 10 30 50 10 20
Sample Output:

3